+++ main.c
@@ -1,116 +1,9 @@
-#include "raylib.h"
-#include "raymath.h"
-#include "math.h"
-
-#define MAX_TEARS 128
-
-enum Direction
-{
-	UP,
-	RIGHT,
-	DOWN,
-	LEFT
-};
-
-typedef enum
-{
-	WALL,
-	FIRE,
-	POOP,
-	EMPTY,
-	DOOR
-} TileType;
-
-typedef union
-{
-	struct
-	{
-		int hitsLeft;
-		int damage;
-
-	} fire;
-	struct
-	{
-		int hitsLeft;
-	} poop;
-
-} TileData;
-
-typedef struct
-{
-	Rectangle rect;
-	Color color;
-	bool isSolid;
-	TileType type;
-	TileData data;
-} Tile;
-
-typedef struct
-{
-	Vector2 position;
-	Vector2 startPosition;
-	enum Direction direction;
-	bool exists;
-} Tear;
-
-typedef struct
-{
-	int items[MAX_TEARS];
-	int head;
-	int tail;
-} TearQueue;
-
-void initTearQueue(TearQueue *q)
-{
-	q->head = -1;
-	q->tail = 0;
-}
-
-void enqueueTear(TearQueue *q, int val)
-{
-	q->items[q->tail] = val;
-	q->tail = (q->tail + 1) % MAX_TEARS;
-
-	if (q->tail == q->head)
-	{
-		q->head = (q->head + 1) % MAX_TEARS;
-	}
-}
-
-int dequeueTear(TearQueue *q)
-{
-	q->head = (q->head + 1) % MAX_TEARS;
-	int val = q->items[q->head];
-	return val;
-}
-
-void spawnTear(Tear *tears, Vector2 pos, enum Direction dir, TearQueue *q)
-{
-	int found = 0;
-	int aux;
-	for (int i = 0; i < MAX_TEARS; ++i)
-	{
-		if (!tears[i].exists)
-		{
-			tears[i].position = pos;
-			tears[i].startPosition = pos;
-			tears[i].direction = dir;
-			tears[i].exists = true;
-			found = 1;
-			enqueueTear(q, i);
-			break;
-		}
-	}
-	if (!found)
-	{
-		aux = dequeueTear(q);
-		tears[aux].position = pos;
-		tears[aux].startPosition = pos;
-		tears[aux].direction = dir;
-		enqueueTear(q, aux);
-	}
-}
-// TODO: Add different types of tiles
+#include <math.h>
+#include <raylib.h>
+#include <raymath.h>
+#include "tear.h"
+#include "tile.h"
+#include "movement.h"
 
 int main(void)
 {
