+++ movement.h
@@ -1,0 +1,12 @@
+#ifndef MOVEMENT_H
+#define MOVEMENT_H
+
+enum Direction
+{
+	UP,
+	RIGHT,
+	DOWN,
+	LEFT
+};
+
+#endif
